# api/openapi.yaml
openapi: 3.1.0
info:
  title: Stadtfest-Finder API
  version: 0.1.0
  summary: REST API for the Stadtfest-Finder mobile app.
  description: |
    Single source of truth for the REST interface (API-first, see CLAUDE.md).
    Times are ISO 8601; "today" is evaluated in Europe/Berlin. Coordinates are WGS84.
    Public endpoints work without a token; protected ones require a JWT issued by the IdP.
  contact:
    name: Stadtfest-Finder
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /v1/geocode:
    get:
      operationId: geocode
      tags: [geocoding]
      summary: Search places and ZIP codes in Germany
      description: |
        Suggestions for a city name, ZIP code or address (Germany only). A 5-digit input is
        treated as a ZIP code. Rate-limited per client.
      security: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search text (city, ZIP code or address).
          schema:
            type: string
            minLength: 2
            maxLength: 120
        - name: limit
          in: query
          required: false
          description: Maximum number of suggestions.
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 5
      responses:
        "200":
          description: Suggestions, best match first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeocodeResult"
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /v1/geocode/reverse:
    get:
      operationId: reverseGeocode
      tags: [geocoding]
      summary: Place for coordinates
      description: Resolves coordinates to ZIP code and place. Coordinates are never stored.
      security: []
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitude (WGS84).
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: lon
          in: query
          required: true
          description: Longitude (WGS84).
          schema:
            type: number
            minimum: -180
            maximum: 180
      responses:
        "200":
          description: The place at the coordinates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReverseGeocodeResult"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/mod/geocode/reverse:
    get:
      operationId: reverseGeocodeEventLocation
      tags: [moderation]
      summary: Address of an event pin
      description: |
        Resolves the pin of an event to street, ZIP code and place (moderators only). Unlike
        `/v1/geocode/reverse` the coordinate is used in full precision: an event location is
        public data, not a user position. `street` is missing if no address is near the pin.
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitude (WGS84).
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: lon
          in: query
          required: true
          description: Longitude (WGS84).
          schema:
            type: number
            minimum: -180
            maximum: 180
      responses:
        "200":
          description: The place at the pin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReverseGeocodeResult"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by the IdP (Zitadel in production, Keycloak locally).
  responses:
    Error:
      description: Error in the common error format.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    GeocodeResult:
      type: object
      description: A geocoding suggestion.
      required: [label, city, lat, lon, kind]
      properties:
        label:
          type: string
          examples: ["73430 Aalen"]
        postalCode:
          type: ["string", "null"]
          pattern: "^[0-9]{5}$"
        city:
          type: string
        lat:
          type: number
        lon:
          type: number
        kind:
          type: string
          enum: [postcode, city, address]
    ReverseGeocodeResult:
      type: object
      description: Place for coordinates.
      required: [city, label]
      properties:
        street:
          type: ["string", "null"]
          description: Street and house number; only `/v1/mod/geocode/reverse` fills it.
        postalCode:
          type: ["string", "null"]
          pattern: "^[0-9]{5}$"
        city:
          type: string
        label:
          type: string
...
    Error:
      type: object
      description: Common error format for all non-2xx responses.
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code, e.g. `validation_failed`, `not_found`.
          examples: [validation_failed]
        message:
          type: string
          description: Human-readable message (German, suitable for display).
          examples: [Bitte fülle die markierten Pflichtfelder aus]
        fields:
          type: object
          description: Field-level errors keyed by field name.
          additionalProperties:
            type: string
...
security:
  - bearerAuth: []